% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dnnCvC}
\alias{dnnCvC}
\title{DNN Classifier with Cross-Validation}
\usage{
dnnCvC(train_data, train_labels, s_values)
}
\arguments{
\item{train_data}{A numeric matrix where rows are training samples and columns are features}

\item{train_labels}{An integer vector representing labels for the training data}

\item{s_values}{A vector of integers representing the candidate values of s for cross-validation}
}
\value{
The optimal s value that maximizes the average cross-validation accuracy
}
\description{
A DNN implementation that selects the best s using 5-fold cross-validation
}
\examples{
\dontrun{
train_data <- matrix(rnorm(200 * 2), ncol = 2)
train_labels <- sample(c(0, 1), size = 200, replace = TRUE)
s_values <- c(1, 2, 3, 4, 5)
best_s <- dnnCvC(train_data, train_labels, s_values)
print(best_s)
}
}
