% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbs_sampler}
\alias{gibbs_sampler}
\title{Gibbs Sampler for a Binomial-Beta Model}
\usage{
gibbs_sampler(n_iter, n, a, b)
}
\arguments{
\item{n_iter}{An integer specifying the number of iterations for the Gibbs sampler.}

\item{n}{An integer representing the total number of trials for the Binomial distribution.}

\item{a}{A numeric value specifying the shape parameter \code{a} of the Beta prior.}

\item{b}{A numeric value specifying the shape parameter \code{b} of the Beta prior.}
}
\value{
A numeric matrix of size \code{n_iter} x 2 where each row contains a sample of \code{x} (column 1) and \code{y} (column 2).
}
\description{
This function performs Gibbs sampling for a Binomial-Beta hierarchical model. 
It generates samples from the joint posterior distribution of \code{x} and \code{y}.
}
\details{
The Gibbs sampler alternates between two steps:
\itemize{
  \item Sampling \code{x} given \code{y} from a Binomial distribution with parameters \code{n} and \code{y}.
  \item Sampling \code{y} given \code{x} from a Beta distribution with parameters \code{x + a} and \code{n - x + b}.
}
This procedure approximates the joint posterior distribution of \code{x} and \code{y}.
}
\examples{
\dontrun{
n_iter <- 1000
n <- 10
a <- 2
b <- 2
samples <- gibbs_sampler(n_iter, n, a, b)
plot(samples[,1], samples[,2], main = "Gibbs Sampler", xlab = "x", ylab = "y")
}
}
