% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{knnCvC}
\alias{knnCvC}
\title{K-Nearest Neighbors Classifier with Cross-Validation}
\usage{
knnCvC(train_data, train_labels, k_values)
}
\arguments{
\item{train_data}{A numeric matrix where rows are training samples and columns are features}

\item{train_labels}{An integer vector representing labels for the training data}

\item{k_values}{A vector of integers representing the candidate values of k for cross-validation}
}
\value{
The optimal k value that maximizes the average cross-validation accuracy
}
\description{
A KNN implementation that selects the best K using 5-fold cross-validation
}
\examples{
\dontrun{
train_data <- matrix(rnorm(200 * 2), ncol = 2)
train_labels <- sample(c(0, 1), size = 200, replace = TRUE)
k_values <- c(1, 2,3,4,5)
best_k <- knnCvC(train_data, train_labels, k_values)
print(best_k)
}
}
