% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homework.R
\name{Required R Packages}
\alias{Required R Packages}
\title{Required R Packages}
\description{
The following R packages are required for running the examples and analyses in the vignette: microbenchmark ,Rcpp,ggplot2,xtable,boot,stats
}
