\name{SA24204158-package}
\alias{SA24204158-package}
\alias{SA24204158}
\docType{package}
\title{
  Distance-Based Classifiers
}
\description{
  This package provides implementations of various distance-based classifiers, including 
  K-Nearest Neighbors (KNN), Distributed Nearest Neighbors (DNN), and Two-Scale Distributed 
  Nearest Neighbors (TDNN). The package leverages both R and C++ implementations for optimized 
  performance and offers flexible solutions for classification tasks.
}
\author{
  SA24204158 <745036240@qq.com>
}
\keyword{ package }

